pyproject.toml
[tool.poetry]
name = "yandex-music-rpc"
version = "0.1.0"
description = ""
authors = ["Aleksandr <waitstop> Kazakov"]
readme = "README.md"
packages = [{include = "yandex_music_rpc"}]

[tool.poetry.dependencies]
python = "^3.10"
customtkinter = "^5.2.0"
configparser = "^6.0.0"
pillow = "^10.0.0"
pystray = "^0.19.4"
yandex-music = "^2.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
